edit env file
