+++ config/default.js
@@ -1,6 +1,6 @@
 const PORT = process.env.PORT ? process.env.PORT : 3001;
-const DBURL = process.env.DBURL ? process.env.DBURL : '';
-const JWTKEY = process.env.JWTKEY ? process.env.JWTKEY : '';
+const DBURL = process.env.DBURL ? process.env.DBURL : 'mongodb://localhost:27017';
+const JWTKEY = process.env.JWTKEY ? process.env.JWTKEY : Math.random().toString();
 
 module.exports = {
 	PORT,
