+++ utils/auth.js
@@ -2,8 +2,6 @@ const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
 const { JWTKEY } = require('../config/default');
 
-
-//token generating
 function generateJWt(payload) {
 	return jwt.sign(payload, JWTKEY, { expiresIn: '24h' })
 }
@@ -18,25 +16,13 @@ async function verifyJwt(token, secret) {
 	}
 }
 
-
-
-
-// export async function generateToken() {
-// 	const token = await TokenGenerator({
-// 		salt: TSALT,
-// 		timestampMap: TIMESTAMPMAP,
-// 	}).generate();
-// 	return token
-// }
-
-//hash pass
 async function hashPass(textPass) {
 	const salt = await bcrypt.genSalt(10);
 	const hashedPass = await bcrypt.hash(textPass, salt);
 	return hashedPass;
 }
 
-//checkPass
+
 async function checkPass(textPass, hashedPass) {
 	try {
 		return !!(await bcrypt.compare(textPass, hashedPass))
