+++ views/layout.pug
@@ -26,8 +26,6 @@ mixin login-template(idButton, nameButton)
     div.hidden#notification-block
       p#notification.form__notification= ''
 
-
-
 html
   head
     link(rel="stylesheet" href="/public/css/style.css")
